Hacky fixes for using enums in imported projects
Protobuf.NET has some issues with referring to enums which are defined in
separate .proto files which have been imported.  The real fix to this
probably involves some fairly extensive changes to the .xslt template, but
as a simple workaround I've put in the following hacky solution:

1. First, we parse the .proto file into a FileDescriptorProto as usual.
2. We loop through all the listed dependencies and merge those with the
   current FileDescriptorProto.  We do this recursively.
3. This brings all defined enums into scope.  We then take advantage of
   the fact that the file we originally wanted always comes first, by
   scanning through the generated .cs file for the start of the second
   output .proto file, and deleting everything from that point on.

In theory, we should have been able to use the -d flag to Protobuf.NET to
achieve this without making any modifications, but that flag seems to be
a little buggy in the way that it renders enums, so I've put in this hack
as a workaround.